+++ src/test/lib/clients/aizu_online_judge.test.ts
@@ -2,8 +2,9 @@ import { describe, test, expect } from 'vitest';
 
 import { loadMockData } from '../common/test_helpers';
 
-import { ContestSiteApiClient } from '$lib/clients/common';
+import { ContestSiteApiClient } from '$lib/clients/http_client';
 import { AojApiClient } from '$lib/clients/aizu_online_judge';
+
 import type { ContestsForImport } from '$lib/types/contest';
 import type { TasksForImport } from '$lib/types/task';
